+++ src/cpm/mod.rs
@@ -50,6 +50,7 @@ pub mod dazzler;
 pub mod fetch;
 pub mod printer;
 pub mod screen;
+pub mod speed;
 pub mod tarbell;
 pub mod uart;
 pub mod vdm;